feat: ensure consistent tree instances for behaviors
Previously, when a behavior is applied to a projection node, the instances returned from tree query methods such as `parent()` `children()` still return the inner type.
This is problematic, since when traversing the tree, we only have access to the basic projection node, which leads in-perfect animations, and also the metadata are lost.
This commit ensures these methods return also the behavior instances by dynamically decorating the return values from the inner instance.
In order to ensure we do not create multiple duplicate behavior instances for a single inner node, this commit forbids the direct use of behavior constructor and forces static methods for creating behaviors, which maintains a weak map of instances to make sure there is only one instance for each inner node.